Factor out a common `TCPDevice` class.
